+++ views/diagnostics-page.php
@@ -170,36 +170,4 @@ if ( ! defined( 'ABSPATH' ) ) {
 
         </tbody>
     </table>
-
-    <?php
-    if ( $view['memcachedStats'] ?? false ) : ?>
-    <h4>Memcached Stats</h4>
-
-        <?php foreach ( $view['memcachedStats'] as $server => $stats ) : ?>
-
-    <h5>Server: <?php echo esc_html( $server ); ?></h5>
-
-    <table class="widefat striped">
-        <thead>
-            <tr>
-                <th>Name</th>
-                <th>Value</th>
-            </tr>
-        </thead>
-        <tbody>
-
-            <?php foreach ( $stats as $name => $value ) : ?>
-            <tr>
-            <td><?php echo esc_html( $name ); ?></td>
-            <td><?php echo esc_html( $value ); ?></td>
-            </tr>
-
-            <?php endforeach; ?>
-
-        </tbody>
-    </table>
-
-    <?php endforeach; ?>
-
-    <?php endif; ?>
 </div>